+++ src/common.h
@@ -1,5 +1,14 @@
 #pragma once
 #ifdef _WIN32
+inline constexpr const char *SHOWPOS_PLATFORM = "windows";
+inline constexpr const char *SHOWPOS_PLATFORM_LABEL = "Windows x64";
+inline constexpr const char *SHOWPOS_SERVER_MODULE = "server.dll";
+#else
+inline constexpr const char *SHOWPOS_PLATFORM = "linux";
+inline constexpr const char *SHOWPOS_PLATFORM_LABEL = "Linux x86_64";
+inline constexpr const char *SHOWPOS_SERVER_MODULE = "libserver.so";
+#endif
+#ifdef _WIN32
 #pragma warning(disable: 4146 4244 4267 4018 4099 4005 5033)
 #endif
 #include <ISmmPlugin.h>